+++ nlisim/modules/hemoglobin.py
@@ -86,7 +86,7 @@ class Hemoglobin(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(hemoglobin.grid) / voxel_volume,
+            'concentration': float(np.mean(hemoglobin.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
