+++ nlisim/modules/il10.py
@@ -110,7 +110,7 @@ class IL10(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(il10.grid) / voxel_volume,
+            'concentration': float(np.mean(il10.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
